feat: remove UI from production bundle
