+++ package.json
@@ -3,52 +3,23 @@
   "description": "Split testing tool for the Web",
   "version": "4.1.0",
   "author": "Trustpilot A/S",
-  "bugs": {
-    "url": "https://github.com/trustpilot/skift/issues"
-  },
-  "dependencies": {},
-  "devDependencies": {
-    "@commitlint/cli": "7.0.0",
-    "@commitlint/config-conventional": "7.0.1",
-    "@semantic-release/git": "6.0.1",
-    "@types/jest": "23.1.4",
-    "ajv": "6.5.2",
-    "css-loader": "0.28.11",
-    "css-to-string-loader": "0.1.3",
-    "html-webpack-plugin": "3.2.0",
-    "husky": "0.14.3",
-    "jest": "23.2.0",
-    "jquery": "3.3.1",
-    "querystringify": "2.0.0",
-    "rimraf": "2.6.2",
-    "semantic-release": "15.6.1",
-    "svg-inline-loader": "0.8.0",
-    "ts-jest": "23.0.0",
-    "ts-loader": "4.4.2",
-    "tslint": "5.10.0",
-    "typescript": "2.9.2",
-    "uglifyjs-webpack-plugin": "1.2.7",
-    "webpack": "4.14.0",
-    "webpack-cli": "3.0.8",
-    "webpack-serve": "1.0.4"
-  },
-  "files": [
-    "dist",
-    "src",
-    "types"
-  ],
   "homepage": "https://github.com/trustpilot/skift#readme",
   "license": "MIT",
-  "main": "dist/skift.js",
+  "main": "lib/index.js",
+  "types": "types/index.d.ts",
   "repository": {
     "type": "git",
     "url": "https://github.com/trustpilot/skift.git"
   },
+  "bugs": {
+    "url": "https://github.com/trustpilot/skift/issues"
+  },
   "scripts": {
-    "build": "npm run clean && npm run build:dev && npm run build:prod",
+    "build": "npm run clean && npm run build:dev && npm run build:ts && npm run build:prod",
     "build:dev": "webpack-cli --config ./config/webpack.dev.js",
     "build:prod": "webpack-cli --config ./config/webpack.prod.js",
-    "clean": "rimraf dist/ yarn-error.log npm-debug.log",
+    "build:ts": "tsc --project .",
+    "clean": "rimraf lib/ dist/ types/ yarn-error.log npm-debug.log",
     "commitmsg": "commitlint -E GIT_PARAMS",
     "lint": "tslint 'src/**/*.ts' --project ./",
     "lint:fix": "tslint 'src/**/*.ts' --project ./ --fix",
@@ -56,7 +27,39 @@
     "test": "jest --config ./config/jest.config.js",
     "ts": "tsc"
   },
-  "types": "types/index.d.ts",
+  "dependencies": {
+    "querystringify": "2.0.0"
+  },
+  "devDependencies": {
+    "@commitlint/cli": "7.0.0",
+    "@commitlint/config-conventional": "7.0.1",
+    "@semantic-release/git": "6.0.1",
+    "@types/jest": "23.1.4",
+    "@types/node": "10.5.1",
+    "ajv": "6.5.2",
+    "css-loader": "0.28.11",
+    "css-to-string-loader": "0.1.3",
+    "html-webpack-plugin": "3.2.0",
+    "husky": "0.14.3",
+    "jest": "23.3.0",
+    "rimraf": "2.6.2",
+    "semantic-release": "15.6.3",
+    "svg-inline-loader": "0.8.0",
+    "ts-jest": "23.0.0",
+    "ts-loader": "4.4.2",
+    "tslint": "5.10.0",
+    "typescript": "2.9.2",
+    "uglifyjs-webpack-plugin": "1.2.7",
+    "webpack": "4.15.0",
+    "webpack-cli": "3.0.8",
+    "webpack-serve": "1.0.4"
+  },
+  "files": [
+    "dist",
+    "src",
+    "types",
+    "lib"
+  ],
   "release": {
     "prepare": [
       "@semantic-release/npm",
